split shorten path into its own file
